bugfix: round time when calling timeout_add
Avoids a warning when the time is a float

+++ papyon/util/timer.py
@@ -37,7 +37,7 @@ class Timer(object):
         if int(time) == time:
             source = gobject.timeout_add_seconds(int(time), self.on_timeout, key)
         else:
-            source = gobject.timeout_add(time * 1000, self.on_timeout, key)
+            source = gobject.timeout_add(int(time * 1000), self.on_timeout, key)
         self._timeout_sources[key] = source
         self._timeout_args[key] = cb_args
